tui/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lookout-tui"
version = "0.1.0"
description = "Terminal UI for Lookout governance-as-code platform"
requires-python = ">=3.10"
license = "Elastic-2.0"
dependencies = [
    "lookout",
    "textual>=0.85.0",
    "litellm>=1.40.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
lookout-tui = "lookout_tui.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lookout_tui"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
strict = true
packages = ["lookout_tui"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["lookout.*", "textual.*", "litellm", "litellm.*"]
ignore_missing_imports = true
